♻️ refactor: add header directory
