+++ src/components/header/Header.tsx
@@ -1,12 +1,12 @@
 'use client';
 import { usePathname } from 'next/navigation';
 import Image from 'next/image';
-import Button from './Button';
+import Button from '../Button';
 import UserInfo from './UserInfo';
 import styles from './Header.module.css';
 
 interface HeaderProps {
-  component: React.ComponentType;
+  component?: React.ComponentType;
 }
 
 export default function Header({ component: Component }: HeaderProps) {
